ai-news-bot/requirements.txt
python-telegram-bot>=20.7
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.0
psycopg>=3.1.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
apscheduler>=3.10.0 